# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  dtos.CategoryCreateDTO:
    properties:
      name:
        maxLength: 20
        type: string
    required:
    - name
    type: object
  dtos.CategoryReadDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dtos.CategoryUpdateDTO:
    properties:
      id:
        type: integer
      name:
        maxLength: 20
        type: string
    required:
    - id
    - name
    type: object
...
  dtos.ErrResp:
    properties:
      error:
        type: string
      path:
        type: string
      timestamp:
        type: string
    type: object
...
  dtos.UserCreateDTO:
    properties:
      email:
        maxLength: 30
        type: string
      password:
        maxLength: 50
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dtos.UserLoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dtos.UserReadDTO:
    properties:
      banned_until:
        type: string
      email:
        type: string
      id:
        type: integer
      is_banned:
        type: boolean
      last_login:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: This site documents the endpoints for the smaash web app, allowing
    easy testing
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: SMAASH API documentation
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user
      parameters:
      - description: dto for logging in a user
        in: body
        name: user_login_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.UserLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: returns the id of the logged in user
          schema:
            type: int
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "422":
          description: request body in wrong format
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - auth
...
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: dto for creating a new user
        in: body
        name: user_create_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.UserCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: returns newly created user
          schema:
            $ref: '#/definitions/dtos.UserReadDTO'
        "409":
          description: unique key violation
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "422":
          description: request body in wrong format
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - auth
  /categories:
    get:
      consumes:
      - application/json
      description: Reads categories
      produces:
      - application/json
      responses:
        "200":
          description: returns all categories
          schema:
            items:
              $ref: '#/definitions/dtos.CategoryReadDTO'
            type: array
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category
      parameters:
      - description: dto for creating a new category
        in: body
        name: category_create_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: returns newly created category
          schema:
            $ref: '#/definitions/dtos.CategoryReadDTO'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "409":
          description: unique key violation
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "422":
          description: request body in wrong format
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a category with the given id
      parameters:
      - description: id of desired category
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: doesn't return anything
          schema:
            type: ""
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Reads a category by it's id
      parameters:
      - description: the id of the desired category
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns the desired category
          schema:
            $ref: '#/definitions/dtos.CategoryReadDTO'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Updates the category with the given id
      parameters:
      - description: dto for updating a category
        in: body
        name: category_update_dto
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryUpdateDTO'
      - description: id of desired category
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: doesn't return anything
          schema:
            type: ""
        "400":
          description: id from url and id from request body doesn't match
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "404":
          description: record not found
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "409":
          description: unique key violation
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "422":
          description: request body in wrong format
          schema:
            $ref: '#/definitions/dtos.ErrResp'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/dtos.ErrResp'
      tags:
      - categories
...
swagger: "2.0"